Fix preference data type to float

+++ db/migrate/20171028135737_create_user_category_preference.rb
@@ -3,7 +3,7 @@ class CreateUserCategoryPreference < ActiveRecord::Migration[5.1]
     create_table :user_category_preference, id: false do |t|
       t.references :category, foreign_key: true
       t.references :user, foreign_key: true
-      t.integer :preference
+      t.float :preference
 
       t.timestamps
     end
